Better fix for race condition.
Explanation:

Interactions (may) push newly generated redexes. This can happen in move()
directly, move() via link(), and expand_ref().

It is my observation that if one of those iteraction-generated redexes is
stolen and processed before the completion of all subsequent links and moves
which occur within an interaction, that things go sideways.

In other words, it is my hypothesis that an interation must be "complete"
before any of the redexes it generates are processed.

The previous commit disabled pushing redexes to the "booty bag" (the bag of
redexes which can be stolen) while processing an interaction. That seems to
work, but it also may have slowed things down a bit; I observe the best perf
when redexes are pushed to the booty bag as aggressively as possible.

This commit changes the logic a bit: allow pushing redexes to the booty bag
while processing an interaction, but if the booty bag gets filled, delay
marking it full until after the interaction is complete, thus preventing it
from being stolen, and ensuring any redexes it contains will not be processed
while the interaction that generated them is still active.